add back cy.checkColorContrast()
