+++ cypress/support/commands.ts
@@ -52,3 +52,20 @@ Cypress.Commands.add("checkAccessibility", () => {
     cy.injectAxe();
     cy.checkA11y(undefined, undefined, terminalLog);
 });
+
+Cypress.Commands.add("checkColorContrast", () => {
+    const terminalLog = (violations: Result[]): void => {
+        cy.task(
+            "table",
+            violations.map(({ id, impact, description, nodes }) => ({
+                id,
+                impact,
+                description,
+                length: nodes.length,
+            }))
+        );
+    };
+
+    cy.injectAxe();
+    cy.checkA11y(undefined, { runOnly: { type: "tag", values: ["wcag2aa"] } }, terminalLog);
+});
